refactor: address daily quality findings for 2026-08-04
- DriveResult err/ok helpers + enrich_drive_success in desktop-use control
- Split drive_tools clipboard/app lifecycle builders
- Split CLI desktop-use into args/drive/mod modules
- Extract webview runtime inject and guest color-scheme from surface-manager
- Extract BrowserSession i18n helper (browserT)
- Add quality review result for 2026-08-04 (score 77)
